+++ components/WorkExperience.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function WorkExperience() {
+  return (
+    <div className='w-full sectionPadding'>
+        <div className="">
+            <h1 className="sectionOverline">work experience</h1>
+            <h1 className="sectionHeader">what have i done?</h1>
+        </div>
+    </div>
+  )
+}
